pyproject.toml
[tool.poetry]
name = "until-zero"
version = "0.1.0"
description = ""
authors = ["u8slvn <u8slvn@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "until_zero"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"
playsound = "1.2.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
mypy = "^1.4.1"
types-pillow = "^10.0.0.1"
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
types-playsound = "^1.3.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 99

[tool.ruff]
fix = true
line-length = 99
extend-select = [
    "I",   # isort
]

[tool.ruff.isort]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2
required-imports = ["from __future__ import annotations"]

[tool.coverage.report]
exclude_lines = [
    "pass",
    "if TYPE_CHECKING:",
]
